Refactor: metadata field validators
